pyproject.toml
[tool.poetry]
name = "minecraft-server-discord-status"
version = "0.1.0"
description = "Updates a Discord message with the status of a Minecraft server"
authors = ["Joel <joel@joel.tokyo>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.7"
"discord.py" = "^1.7.3"
mcstatus = "^9.0.4"

[tool.poetry.dev-dependencies]
black = "^22.1.0"
flake8 = "^4.0.1"
isort = "^5.10.1"
mypy = "^0.931"
pytest = "^7.0.1"
pytest-cov = "^3.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"
lines_after_imports = 2

[tool.mypy]
install_types = true
ignore_missing_imports = true
non_interactive = true
strict = true
